Type cast size_t to int as needed. size_t is U64 on windows 64 bit machines. GCC defines size_t as int. It is reasonable to type definitions as size_t when that are relating to the sizeof() elements. For portability with windows casting size_t to int should be alright since most of the left-sided elements are usually int.
git-svn-id: file:///home/jj/hercules.svn/trunk@6021 956126f8-22a0-4046-8f4a-272fa8102e63

+++ general1.c
@@ -4042,12 +4042,12 @@ size_t  dstlen,srclen;                  /* Page wide src/dst lengths */
         BYTE *source;
         /* get source frame and copy concurrently */
         source = MADDR (addr2, r3, regs, ACCTYPE_READ, regs->psw.pkey);
-        concpy(regs,dest,source,copylen);
+        concpy(regs,dest,source,(int)copylen);
         /* Adjust operands */
-        addr2+=copylen;
-        len2-=copylen;
-        addr1+=copylen;
-        len1-=copylen;
+        addr2+=(int)copylen;
+        len2-=(int)copylen;
+        addr1+=(int)copylen;
+        len1-=(int)copylen;
 
         /* Adjust length & pointers for this cycle */
         dest+=copylen;
@@ -4060,8 +4060,8 @@ size_t  dstlen,srclen;                  /* Page wide src/dst lengths */
         memset(dest,pad,dstlen);
 
         /* Adjust destination operands */
-        addr1+=dstlen;
-        len1-=dstlen;
+        addr1+=(int)dstlen;
+        len1-=(int)dstlen;
     }
 
     /* Update the registers */
